+++ src/backends/raylib/raylibwindow.h
@@ -39,7 +39,9 @@ namespace SunLight  {
 
                 public:
 
-                bool Create( int nWidth, int nHeight, const char *szTitle, bool bResizeable ) override;
+                bool Create( int nWidth, int nHeight, const char *szTitle, bool bResizeable,
+                             bool bFullscreen = false,
+                             SunLight :: Window :: FullscreenStrategy strategy = SunLight :: Window :: FULLSCREEN_STRATEGY_REAL ) override;
                 void Close( void ) override;
                 bool ShouldClose( void ) override;
                 void SetExitKey( SunLight :: Input :: KeyboardKey key ) override;
@@ -52,6 +54,7 @@ namespace SunLight  {
                                     SunLight :: Window :: FullscreenStrategy strategy =
                                         SunLight :: Window :: FULLSCREEN_STRATEGY_REAL ) override;
                 bool GetFullscreen( void ) override;
+                SunLight :: Window :: FullscreenStrategy GetFullscreenStrategy( void ) override;
 
                 double GetElapsedTime( void ) override;